% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.clean_match_advanced_stats_data}
\alias{.clean_match_advanced_stats_data}
\title{Clean each match advanced statistic tables}
\usage{
.clean_match_advanced_stats_data(df_in)
}
\arguments{
\item{df_in}{a raw match stats data frame}
}
\value{
a cleaned data frame for the selected match advanced statistic
}
\description{
Returns cleaned data frame for each of the team statistic tables for each selected match
}
