% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.clean_player_season_stats}
\alias{.clean_player_season_stats}
\title{Clean player season statistic tables}
\usage{
.clean_player_season_stats(input_table_element)
}
\arguments{
\item{input_table_element}{element of the html table on the player page}
}
\value{
a data frame for the selected player's advanced statistic
}
\description{
Returns cleaned dataframe for each of the player statistic tables used by fb_player_season_stats()
}
