% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.clean_table_names}
\alias{.clean_table_names}
\title{Clean stat table column names}
\usage{
.clean_table_names(df_in)
}
\arguments{
\item{df_in}{a raw match stats data frame}
}
\value{
a data frame with cleaned names
}
\description{
Returns cleaned column names for stats tables
}
