% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.convert_value_to_numeric}
\alias{.convert_value_to_numeric}
\title{Convert formatted valuations to numeric}
\usage{
.convert_value_to_numeric(euro_value)
}
\arguments{
\item{euro_value}{raw valuation from transfermarkt.com}
}
\value{
a cleaned numeric data value for market and/or transfer valuation
}
\description{
Returns a numeric data type for player valuations
}
