% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals_loading.R
\name{.file_reader}
\alias{.file_reader}
\title{Read in stored RDS data}
\usage{
.file_reader(file_url)
}
\arguments{
\item{file_url}{URL to RDS file(s) for reading in}
}
\value{
data type dependent on RDS file being read in
}
\description{
Reads in RDS stored data files typically stored on GitHub
}
