% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.get_clean_understat_json}
\alias{.get_clean_understat_json}
\title{Clean Understat JSON data}
\usage{
.get_clean_understat_json(page_url, script_name)
}
\arguments{
\item{page_url}{understat.com page URL}

\item{script_name}{html JSON script name}
}
\value{
a cleaned Understat data frame
}
\description{
Returns a cleaned Understat data frame
}
