% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_results.R
\name{.get_each_season_results}
\alias{.get_each_season_results}
\title{Get single match results}
\usage{
.get_each_season_results(fixture_url, time_pause = 2)
}
\arguments{
\item{fixture_url}{URL of the league season's fixtures and results page}

\item{time_pause}{the wait time (in seconds) between page loads}
}
\value{
returns a raw dataframe with the results of the league season
}
\description{
Internal function used in get_match_results
}
\examples{
\donttest{
df <- .get_each_season_results(fixture_url =
"https://fbref.com/en/comps/9/schedule/Premier-League-Scores-and-Fixtures"
)

}
}
