% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.get_league_season_url}
\alias{.get_league_season_url}
\title{Get URL of league season}
\usage{
.get_league_season_url(country, gender, season)
}
\arguments{
\item{country}{the three character country code}

\item{gender}{gender of competition, either "M" or "F"}

\item{season}{the year the season concludes, in quotes, ie "2021"}
}
\value{
a URL for the selected league seasons
}
\description{
Returns a URL for the selected league seasons
}
