% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_report.R
\name{.get_match_report_page}
\alias{.get_match_report_page}
\title{Get match report for webpage}
\usage{
.get_match_report_page(match_page)
}
\arguments{
\item{match_page}{the fbref.com URL for the required match}
}
\value{
returns a dataframe with the match details for a selected match
}
\description{
Internal function to return match report details for a selected match, with input being a webpage, not url
}
