% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.get_tier1_competitions}
\alias{.get_tier1_competitions}
\title{Get Tier 1 competitions}
\usage{
.get_tier1_competitions()
}
\value{
returns a dataframe with the tier 1 competitions from around the world
}
\description{
Returns a df of the top leagues around the world
}
