% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.replace_empty_na}
\alias{.replace_empty_na}
\title{Replace Empty Values}
\usage{
.replace_empty_na(val)
}
\arguments{
\item{val}{a value that can either be empty, or not empty}
}
\value{
NA_character where the extracted value is empty, or the value itself
}
\description{
Returns a NA character for empty values
}
