% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.tm_fix_dates}
\alias{.tm_fix_dates}
\title{Clean date fields}
\usage{
.tm_fix_dates(dirty_dates)
}
\arguments{
\item{dirty_dates}{formatted date value}
}
\value{
a cleaned date
}
\description{
Returns a date format in YYYY-MM-DD from 'mmm d, yyyy'
}
