% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.understat_shooting}
\alias{.understat_shooting}
\title{Understat shots location helper function}
\usage{
.understat_shooting(type_url)
}
\arguments{
\item{type_url}{can be season, team, match, player URL}
}
\value{
a cleaned Understat shooting location data frame
}
\description{
Returns a cleaned Understat shooting location  data frame
}
