% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_league_stats.R
\name{fb_league_stats}
\alias{fb_league_stats}
\title{Get FBref Team or Player Season Statistics for an Entire League}
\usage{
fb_league_stats(
  country,
  gender,
  season_end_year,
  tier = "1st",
  non_dom_league_url = NA,
  stat_type,
  team_or_player,
  time_pause = 3,
  rate = purrr::rate_backoff(max_times = 3)
)
}
\arguments{
\item{country}{the three character country code}

\item{gender}{gender of competition, either "M" or "F", or both}

\item{season_end_year}{the year the season(s) concludes}

\item{tier}{the tier of the league, ie '1st' for the EPL or '2nd' for the Championship and so on}

\item{non_dom_league_url}{the URL for Cups and Competitions found at https://fbref.com/en/comps/}

\item{stat_type}{the type of statistic required. Must be one of the following:
\itemize{
\item{standard}
\item{shooting}
\item{passing}
\item{passing_types}
\item{gca}
\item{defense}
\item{possession}
\item{playing_time}
\item{misc}
\item{keepers}
\item{keepers_adv}
}}

\item{team_or_player}{result either summarised for each team, or individual players}

\item{time_pause}{the wait time (in seconds) between page loads}

\item{rate}{passed to `purrr::insistently`. The function can be brittle, and works only on a second or third try. This parameter controls how much the function will retry to get a result.}
}
\value{
a dataframe of season stats for all teams / players in a league
}
\description{
Get the season stats for all teams / players in a selected league
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
try({
fb_season_team_stats(
  country = "ENG",
  gender = "M",
  season_end_year = 2022,
  tier = "1st",
  stat_type = "shooting",
  team_or_player = "player"
)
## Non-domestic league
##   Note that this is more likely to fail due to the volume of players
fb_league_stats(
  country = NA_character_,
  gender = "M",
  season_end_year = 2023,
  tier = NA_character_,
  non_dom_league_url = "https://fbref.com/en/comps/8/history/Champions-League-Seasons",
  stat_type = "standard",
  team_or_player = "player"
)
})
}
}
