% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldfootballr_helpers.R
\name{fb_league_urls}
\alias{fb_league_urls}
\title{Get fbref League URLs}
\usage{
fb_league_urls(country, gender, season_end_year, tier = "1st")
}
\arguments{
\item{country}{the three character country code}

\item{gender}{gender of competition, either "M" or "F"}

\item{season_end_year}{the year the season(s) concludes (defaults to all available seasons)}

\item{tier}{the tier of the league, ie '1st' for the EPL or '2nd' for the Championship and so on}
}
\value{
returns a character vector of all fbref league URLs for selected country, season, gender and tier
}
\description{
Returns the URLs for season leagues of a selected country
}
\examples{
\dontrun{
fb_league_urls(country = "ENG", gender = "M", season_end_year = 2021, tier = '1st')
}
}
