% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_report.R
\name{fb_match_report}
\alias{fb_match_report}
\title{Get FBref match report}
\usage{
fb_match_report(match_url, time_pause = 3)
}
\arguments{
\item{match_url}{the fbref.com URL for the required match}

\item{time_pause}{the wait time (in seconds) between page loads}
}
\value{
returns a dataframe with the match details for a selected match
}
\description{
Returns match report details for selected matches.
Replaces the deprecated function get_match_report
}
\examples{
\dontrun{
try({
match <- fb_match_urls(country = "AUS", gender = "F", season_end_year = 2021, tier = "1st")[1]
df <- fb_match_report(match_url = match)
})
}
}
