% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_summary.R
\name{fb_match_summary}
\alias{fb_match_summary}
\title{Get FBref match summary}
\usage{
fb_match_summary(match_url, time_pause = 3)
}
\arguments{
\item{match_url}{the fbref.com URL for the required match}

\item{time_pause}{the wait time (in seconds) between page loads}
}
\value{
returns a dataframe with the match events (goals, cards, subs) for selected matches
}
\description{
Returns match summary data for selected match URLs, including goals, subs and cards
Replaces the deprecated function get_match_summary
}
\examples{
\dontrun{
try({
match <- fb_match_urls(country = "AUS", gender = "F", season_end_year = 2021, tier = "1st")[1]
df <- fb_match_summary(match_url = match)
})
}

}
