% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_player_info.R
\name{fb_player_info}
\alias{fb_player_info}
\title{Get Player Attributes}
\usage{
fb_player_info(player_urls, time_pause = 3)
}
\arguments{
\item{player_urls}{the URL(s) of the player(s)}

\item{time_pause}{the wait time (in seconds) between page loads}
}
\value{
A data frame
}
\description{
Return a data frame of a player's info
}
\examples{
\dontrun{
try({
player_urls <- c(
  "https://fbref.com/en/players/d70ce98e/", # Messi
  "https://fbref.com/en/players/3515d404/" # JWP
)
fb_player_info(player_urls)
})
}
}
