% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_player_match_logs.R
\name{fb_player_match_logs}
\alias{fb_player_match_logs}
\title{Get fbref Player Match Logs}
\usage{
fb_player_match_logs(player_url, season_end_year, stat_type)
}
\arguments{
\item{player_url}{the URL of the player (can come from fb_player_urls())}

\item{season_end_year}{the year the season(s) concludes}

\item{stat_type}{the type of statistic required

The statistic type options (stat_type) include:

\emph{"summary"}, \emph{"keepers"}, \emph{"passing"}, \emph{"passing_types"},
\emph{"gca"}, \emph{"defense"}, \emph{"possession"}, \emph{"misc"}}
}
\value{
returns a dataframe of a player's match logs for a season
}
\description{
Returns all match logs for a selected player, season and stat type
}
\examples{
\dontrun{
fb_player_match_logs("https://fbref.com/en/players/3bb7b8b4/Ederson",
season_end_year = 2021, stat_type = 'summary')
}
}
