% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_player_scouting.R
\name{fb_player_scouting_report}
\alias{fb_player_scouting_report}
\title{Get fbref Full Player Scouting Report}
\usage{
fb_player_scouting_report(
  player_url,
  pos_versus,
  league_comp_name = NULL,
  time_pause = 3
)
}
\arguments{
\item{player_url}{the URL of the player (can come from fb_player_urls())}

\item{pos_versus}{either "primary" or "secondary" as fbref offer comparisons against multiple positions}

\item{league_comp_name}{the league or competition name(s) you want the scouting report for. Defaults to all}

\item{time_pause}{the wait time (in seconds) between page loads}
}
\value{
returns a dataframe of a player's full scouting information for all seasons available on FBref
}
\description{
Returns the scouting report for a selected player
}
\examples{
\dontrun{
try({
fb_player_scouting_report(player_url = "https://fbref.com/en/players/d70ce98e/Lionel-Messi",
pos_versus = "primary")

# to filter for the last 365 days:
fb_player_scouting_report(player_url = "https://fbref.com/en/players/d70ce98e/Lionel-Messi",
pos_versus = "primary", league_comp_name = "Last 365 Days Men's Big 5 Leagues, UCL, UEL")

# to get secondary positions
fb_player_scouting_report(player_url = "https://fbref.com/en/players/d70ce98e/Lionel-Messi",
pos_versus = "secondary")

# for the last 365 days and also the 2022 WC
fb_player_scouting_report(player_url = "https://fbref.com/en/players/d70ce98e/Lionel-Messi",
pos_versus = "secondary",
league_comp_name = c("Last 365 Days Men's Big 5 Leagues, UCL, UEL", "2022 World Cup"))
})
}
}
