% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbref_wages.R
\name{fb_squad_wages}
\alias{fb_squad_wages}
\title{Get team player wages}
\usage{
fb_squad_wages(team_urls, time_pause = 3)
}
\arguments{
\item{team_urls}{the URL(s) of the teams(s) (can come from fb_teams_urls())}

\item{time_pause}{the wait time (in seconds) between page loads}
}
\value{
returns a dataframe with all available estimated player wages for the selected team(s)
}
\description{
Returns all player wages from FBref via Capology
}
\examples{
\dontrun{
try({
# for single teams:
man_city_url <- "https://fbref.com/en/squads/b8fd03ef/Manchester-City-Stats"
fb_squad_wages(team_urls = man_city_url)
})
}
}
