% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_team_match_log_stats.R
\name{fb_team_match_log_stats}
\alias{fb_team_match_log_stats}
\title{Get team match log stats}
\usage{
fb_team_match_log_stats(team_urls, stat_type, time_pause = 2)
}
\arguments{
\item{team_urls}{the URL(s) of the teams(s) (can come from fb_teams_urls())}

\item{stat_type}{the type of statistic required}

\item{time_pause}{the wait time (in seconds) between page loads

The statistic type options (stat_type) include:

\emph{"shooting"}, \emph{"keeper"}, \emph{"passing"},
\emph{"passing_types"}, \emph{"gca"}, \emph{"defense"},
\emph{"misc"}}
}
\value{
returns a dataframe with the selected stat outputs of all games played by the selected team(s)
}
\description{
Returns all match statistics for a team(s) in a given season
}
\examples{
\dontrun{
# for single teams:
man_city_url <- "https://fbref.com/en/squads/b8fd03ef/Manchester-City-Stats"
fb_team_match_log_stats(team_urls = man_city_url, stat_type = "passing")
}
}
