% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_match_results.R
\name{fb_team_match_results}
\alias{fb_team_match_results}
\title{Get FBref team match results}
\usage{
fb_team_match_results(team_url, time_pause = 3)
}
\arguments{
\item{team_url}{the URL for the team season}

\item{time_pause}{the wait time (in seconds) between page loads}
}
\value{
returns a dataframe with the results of all games played by the selected team(s)
}
\description{
Returns all game results for a team in a given season
Replaces the deprecated function get_team_match_results
}
\examples{
\dontrun{
try({
# for single teams:
man_city_url <- "https://fbref.com/en/squads/b8fd03ef/Manchester-City-Stats"
fb_team_match_results(man_city_url)
})
}
}
