% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fotmob_leagues.R
\name{fotmob_get_league_ids}
\alias{fotmob_get_league_ids}
\title{Get fotmob league ids}
\usage{
fotmob_get_league_ids(cached = TRUE)
}
\arguments{
\item{cached}{Whether to load the dataframe from the \href{https://github.com/JaseZiv/worldfootballR_data/blob/master/raw-data/fotmob-leagues/all_leagues.csv}{data CSV}. This is faster and most likely what you want to do, unless you identify a league that's being tracked by fotmob that's not in this pre-saved CSV.}
}
\description{
Returns a dataframe of the league ids available on fotmob
}
