% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fotmob_matches.R
\name{fotmob_get_match_momentum}
\alias{fotmob_get_match_momentum}
\title{Get fotmob match momentum}
\usage{
fotmob_get_match_momentum(match_ids)
}
\arguments{
\item{match_ids}{a vector of strings or numbers representing matches}
}
\value{
returns a dataframe of match momentum
}
\description{
Returns match momentum from fotmob.com. Only available for 2022/23 season and beyond.
}
\examples{
\donttest{
try({
match_id <- 3901251
fotmob_get_match_info(match_id)
})
}
}
