% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fotmob_players.R
\name{fotmob_get_match_players}
\alias{fotmob_get_match_players}
\title{Get fotmob match player details by match id}
\usage{
fotmob_get_match_players(match_ids)
}
\arguments{
\item{match_ids}{a vector of strings or numbers representing matches}
}
\value{
returns a dataframe of match players
}
\description{
Returns match details from fotmob.com
}
\examples{
\dontrun{
library(dplyr)
library(tidyr)
## single match
players <- fotmob_get_match_players(3610132)
salah_id <- "292462"
players \%>\%
  dplyr::filter(id == salah_id) \%>\%
  dplyr::select(player_id = id, stats) \%>\%
  tidyr::unnest(stats)

## multiple matches
fotmob_get_match_players(c(3609987, 3609979))
}
}
