% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_lineups.R
\name{get_match_lineups}
\alias{get_match_lineups}
\title{Get match lineups}
\usage{
get_match_lineups(match_url)
}
\arguments{
\item{match_url}{the fbref.com URL for the required match}
}
\value{
returns a dataframe with the team lineups for a selected match
}
\description{
Returns lineups for home and away teams for a selected match
}
\examples{
\dontrun{
match <- "https://fbref.com/en/matches/47880eb7/Liverpool-Manchester-City-November-10-2019-Premier-League"
get_match_lineups(match_url = match)
}
}
