% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_report.R
\name{get_match_report}
\alias{get_match_report}
\title{Get match report}
\usage{
get_match_report(match_url)
}
\arguments{
\item{match_url}{the fbref.com URL for the required match}
}
\value{
returns a dataframe with the match details for a selected match
}
\description{
Returns match report details for a selected match
}
\examples{
\dontrun{
match <- "https://fbref.com/en/matches/47880eb7/Liverpool-Manchester-City-November-10-2019-Premier-League"
get_match_report(match_url = match)
}
}
