% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_results.R
\name{get_match_results}
\alias{get_match_results}
\title{Get match results}
\usage{
get_match_results(country, gender, season_end_year)
}
\arguments{
\item{country}{the three character country code}

\item{gender}{gender of competition, either "M" or "F"}

\item{season_end_year}{the year(s) the season concludes}
}
\value{
returns a dataframe with the results of the competition, season and gender
}
\description{
Returns the game results for a given league season(s)
}
\examples{
\dontrun{
get_match_results(country = c("ENG", "ITA", "GER"), gender = "M", season_end_year = c(2018:2021))
}
}
