% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_shooting.R
\name{get_match_shooting}
\alias{get_match_shooting}
\title{Get match shooting data}
\usage{
get_match_shooting(match_url)
}
\arguments{
\item{match_url}{the fbref.com URL for the required match}
}
\value{
returns a dataframe with detailed player shooting data for a selected match
}
\description{
Returns detailed player shooting data for home and away teams for a selected match
}
\examples{
\dontrun{
match <- "https://fbref.com/en/matches/bf52349b/Fulham-Arsenal-September-12-2020-Premier-League"
get_match_shooting(match_url = match)
}
}
