% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_summary.R
\name{get_match_summary}
\alias{get_match_summary}
\title{Get match summary}
\usage{
get_match_summary(match_url)
}
\arguments{
\item{match_url}{the fbref.com URL for the required match}
}
\value{
returns a dataframe with the match events (goals, cards, subs) for selecetd matches
}
\description{
Returns match summary data for selected match URLs, including goals, subs and cards
}
\examples{
\dontrun{
match <- "https://fbref.com/en/matches/47880eb7/Liverpool-Manchester-City-November-10-2019-Premier-League"
get_match_summary(match_url = match)
}

}
