% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_player_attributes.R
\name{get_player_info}
\alias{get_player_info}
\title{Get Player Attributes}
\usage{
get_player_info(url)
}
\arguments{
\item{url}{URL of the player's profile page on fbref.com}
}
\value{
A data frame
}
\description{
Return a data frame of a player's info
}
