% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_market_values.R
\name{get_player_market_values}
\alias{get_player_market_values}
\title{Get player market values}
\usage{
get_player_market_values(country_name, start_year)
}
\arguments{
\item{country_name}{the country of the league's players}

\item{start_year}{the start year of the season (2020 for the 20/21 season)}
}
\value{
returns a dataframe of player valuations for country/seasons
}
\description{
Returns data frame of player valuations (in Euros) from transfermarkt.com
}
\examples{
\dontrun{
get_player_market_values(country_name = "England", start_year = c(2019, 2020))
}
}
