% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_team_stats.R
\name{get_season_team_stats}
\alias{get_season_team_stats}
\title{Get season team stats}
\usage{
get_season_team_stats(
  country,
  gender,
  season_end_year,
  stat_type = c("league_table", "league_table_home_away", "standard", "keeper",
    "keeper_adv", "shooting", "passing", "goal_shot_creation", "defense", "possession",
    "playing_time", "misc")
)
}
\arguments{
\item{country}{the three character country code}

\item{gender}{gender of competition, either "M" or "F"}

\item{season_end_year}{the year the season concludes, in quotes, ie "2021"}

\item{stat_type}{the type of team statistics the user requires

The statistic type options (stat_type) include:

\emph{"league_table"}, \emph{"league_table_home_away}", \emph{"standard"},
\emph{"keeper"}, \emph{"keeper_adv"}, \emph{"shooting"}, \emph{"passing"},
\emph{"goal_shot_creation"}, \emph{"defense" }, \emph{"possession"},
\emph{"playing_time"}, \emph{"misc"}}
}
\value{
returns a dataframe of a selected team statistic type for a selected league season
}
\description{
Returns different team season statistics results for a given league season and stat type
}
\examples{
\dontrun{
get_team_statistics(country = "ITA", gender = "M", season_end_year = "2020", stat_type = defense)
}
}
