% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_league_suspensions.R
\name{get_suspensions}
\alias{get_suspensions}
\title{Get League Suspensions}
\usage{
get_suspensions(country_name = NA, league_url = NA)
}
\arguments{
\item{country_name}{The country of the league (used if league_url not provided)}

\item{league_url}{Optional direct league URL from transfermarkt.com}
}
\value{
A dataframe of suspended players
}
\description{
Returns a data frame of all suspended players in a selected league
}
