% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_match_results.R
\name{get_team_match_results}
\alias{get_team_match_results}
\title{Get team match results}
\usage{
get_team_match_results(team_url, time_pause = 2)
}
\arguments{
\item{team_url}{the URL for the team season}

\item{time_pause}{the wait time (in seconds) between page loads}
}
\value{
returns a dataframe with the results of all games played by the selected team(s)
}
\description{
Returns all game results for a team in a given season
}
\examples{
\dontrun{
# for single teams:
man_city_url <- "https://fbref.com/en/squads/b8fd03ef/Manchester-City-Stats"
get_team_match_results(man_city_url)
}
}
