% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fb.R
\name{load_fb_advanced_match_stats}
\alias{load_fb_advanced_match_stats}
\title{Load pre-saved FBref match advanced stats}
\usage{
load_fb_advanced_match_stats(
  country,
  gender,
  tier,
  stat_type,
  team_or_player,
  season_end_year = NA
)
}
\arguments{
\item{country}{the three character country code}

\item{gender}{gender of competition, either "M" or "F"}

\item{tier}{the tier of the league, ie '1st' for the EPL or '2nd' for the Championship and so on}

\item{stat_type}{the type of team statistics the user requires}

\item{team_or_player}{result either summarised for each team, or individual players}

\item{season_end_year}{the year(s) the season concludes}
}
\value{
returns a dataframe
}
\description{
Loading version of \code{fb_advanced_match_stats}. Only some leagues available.
}
\examples{
\donttest{
try({
load_fb_advanced_match_stats(
  country = "ENG",
  gender = "M",
  tier = "1st",
  stat_type = "summary",
  team_or_player = "player"
)

load_fb_advanced_match_stats(
  country = c("ITA", "ESP"),
  gender = "M",
  tier = "1st",
  season_end_year = 2023,
  stat_type = "defense",
  team_or_player = "player"
)
})
}
}
