% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fb.R
\name{load_fb_match_summary}
\alias{load_fb_match_summary}
\title{Load pre-saved FBref match summary data}
\usage{
load_fb_match_summary(country, gender, tier, season_end_year = NA)
}
\arguments{
\item{country}{the three character country code}

\item{gender}{gender of competition, either "M" or "F"}

\item{tier}{the tier of the league, ie '1st' for the EPL or '2nd' for the Championship and so on}

\item{season_end_year}{the year(s) the season concludes}
}
\value{
returns a dataframe
}
\description{
Loading version of \code{fb_match_summary}. Only some leagues available.
}
\examples{
\donttest{
try({
load_fb_match_summary(
  country = "ENG",
  gender = "M",
  tier = "1st"
)

load_fb_match_summary(
  country = c("ITA", "ESP"),
  gender = "M",
  tier = "1st",
  season_end_year = 2019
)
})
}
}
