% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fb.R
\name{load_match_comp_results}
\alias{load_match_comp_results}
\title{Load match competition results}
\usage{
load_match_comp_results(comp_name)
}
\arguments{
\item{comp_name}{the three character country code}
}
\value{
returns a dataframe with the results of the competition name
}
\description{
Returns the game results for a competition(s),
ie League cups or international competitions from FBref.
comp_name comes from https://github.com/JaseZiv/worldfootballR_data/tree/master/data/match_results_cups#readme
}
\examples{
\donttest{
try({
df <- load_match_comp_results(
comp_name = "Coppa Italia"
)
# for multiple competitions:
cups <- c("FIFA Women's World Cup",
          "FIFA World Cup")
df <- load_match_comp_results(
comp_name = cups
)
})
}
}
