% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_understat.R
\name{load_understat_league_shots}
\alias{load_understat_league_shots}
\title{Load Understat league shot locations}
\usage{
load_understat_league_shots(league)
}
\arguments{
\item{league}{the available leagues in Understat as outlined below

The leagues currently available for Understat are:
\emph{"EPL"}, \emph{"La liga}", \emph{"Bundesliga"},
\emph{"Serie A"}, \emph{"Ligue 1"}, \emph{"RFPL"}}
}
\value{
returns a dataframe of shooting locations for a selected league
}
\description{
Loading version of \code{understat_league_season_shots}, but for all seasons
Returns shooting locations for all matches played in the selected league
}
\examples{
\dontrun{
try({
  df <- load_understat_league_shots(league = "Serie A")
})
}
}
