% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_dictionary_mapping.R
\name{player_dictionary_mapping}
\alias{player_dictionary_mapping}
\title{Player Mapping Dictionary}
\usage{
player_dictionary_mapping()
}
\value{
returns a dataframe of FBref players and respective Transfermarkt URL
}
\description{
Returns data frame of players from the top 5 Euro leagues, their player URL
and their respective Transfermarkt URL.
Currently only for the players who have been in the top 5 leagues since
the 2017-2018 season
}
\examples{
\donttest{
try({
mapped_players <- player_dictionary_mapping()
})
}
}
