% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_transfer_history.R
\name{player_transfer_history}
\alias{player_transfer_history}
\title{Get player transfer history}
\usage{
player_transfer_history(player_urls)
}
\arguments{
\item{player_urls}{the player url(s) from transfermarkt}
}
\value{
returns a dataframe of player transfers
}
\description{
Returns data frame of player(s) transfer history from transfermarkt.com
}
\examples{
\donttest{
player_transfer_history("https://www.transfermarkt.com/eden-hazard/profil/spieler/50202")
}
}
