% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_transfer_history.R
\name{single_player_transfer_history}
\alias{single_player_transfer_history}
\title{Get player transfer history}
\usage{
single_player_transfer_history(player_url)
}
\arguments{
\item{player_url}{the single player url from transfermarket}
}
\value{
returns a dataframe of player transfers
}
\description{
Returns data frame of player transfer history from transfermarkt.com
Function used as internal function to player_transfer_history()
}
\examples{
\dontrun{
single_player_transfer_history(player_url)
}
}
