% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_expiring_contracts.R
\name{tm_expiring_contracts}
\alias{tm_expiring_contracts}
\title{Get expiring contracts}
\usage{
tm_expiring_contracts(country_name, contract_end_year, league_url = NA)
}
\arguments{
\item{country_name}{the country of the league's players}

\item{contract_end_year}{the year the contract is due to expire}

\item{league_url}{league url from transfermarkt.com. To be used when country_name not available in main function}
}
\value{
returns a dataframe of expiring contracts in the selected league
}
\description{
Returns a data frame of players with expiring contracts for a selected league and time period
}
