% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_player_absence.R
\name{tm_get_player_absence}
\alias{tm_get_player_absence}
\title{Get Player Absences}
\usage{
tm_get_player_absence(player_urls)
}
\arguments{
\item{player_urls}{player url(s) from transfermarkt}
}
\value{
returns a dataframe
}
\description{
Returns data frame of a player's absences from suspension from transfermarkt.com
}
\examples{
\dontrun{
try({
player_urls <- c("https://www.transfermarkt.com/cristian-romero/profil/spieler/355915",
"https://www.transfermarkt.com/micky-van-de-ven/profil/spieler/557459")

df <- tm_get_player_absence(player_urls)
})
}

}
