% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_league_suspensions.R
\name{tm_get_risk_of_suspension}
\alias{tm_get_risk_of_suspension}
\title{Get League Risk of Suspension}
\usage{
tm_get_risk_of_suspension(country_name, league_url = NA)
}
\arguments{
\item{country_name}{The country of the league (used if league_url not provided)}

\item{league_url}{Optional direct league URL from transfermarkt.com}
}
\value{
A data frame of players at risk of suspension
}
\description{
Returns a data frame of players at risk of suspension
}
\examples{
\dontrun{
try({
tm_get_risk_of_suspension(country_name = "England")
tm_get_risk_of_suspension(league_url = "https://www.transfermarkt.com/premier-league/sperrenausfaelle/wettbewerb/GB1")
})
}
}
