% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_league_debutants.R
\name{tm_league_debutants}
\alias{tm_league_debutants}
\title{Get league debutants}
\usage{
tm_league_debutants(
  country_name,
  league_url = NA,
  debut_type,
  debut_start_year,
  debut_end_year
)
}
\arguments{
\item{country_name}{the country of the league's players}

\item{league_url}{league url from transfermarkt.com. To be used when country_name not available in main function}

\item{debut_type}{whether you want 'league' debut or 'pro' debut}

\item{debut_start_year}{the season start year of the beginning of the period you want results for}

\item{debut_end_year}{the season start year of the end of the period you want results for}
}
\value{
returns a dataframe of players who debuted in the selected league
}
\description{
Returns a data frame of debutants for a selected league
}
