% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldfootballr_helpers.R
\name{tm_league_team_urls}
\alias{tm_league_team_urls}
\title{Get transfermarkt Team URLs}
\usage{
tm_league_team_urls(country_name, start_year, league_url = NA)
}
\arguments{
\item{country_name}{the country of the league's players}

\item{start_year}{the start year of the season (2020 for the 20/21 season)}

\item{league_url}{league url from transfermarkt.com. To be used when country_name not avalilable in main function}
}
\value{
returns a character vector of all transfermarkt team URLs for a selected league
}
\description{
Returns the URLs for all teams for a given league season
}
\examples{
\donttest{
team_urls <- tm_league_team_urls(country_name = "England", start_year = 2020)
}
}
