% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_player_bio.R
\name{tm_player_bio}
\alias{tm_player_bio}
\title{Get transfermarkt player bios}
\usage{
tm_player_bio(player_urls)
}
\arguments{
\item{player_urls}{player url(s) from transfermarket}
}
\value{
returns a dataframe of player bios
}
\description{
Returns data frame of player bois from transfermarkt.com
}
\examples{
\dontrun{
player_url <- "https://www.transfermarkt.com/eden-hazard/profil/spieler/50202"
tm_player_bio(player_url)
tm_player_bio(player_urls = c("https://www.transfermarkt.com/eden-hazard/profil/spieler/50202",
                              "https://www.transfermarkt.com/sergio-ramos/profil/spieler/25557",
                              "https://www.transfermarkt.com/ivo-grbic/profil/spieler/226073"))
}
}
