% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_player_injury_history.R
\name{tm_player_injury_history}
\alias{tm_player_injury_history}
\title{Get player injury history}
\usage{
tm_player_injury_history(player_urls)
}
\arguments{
\item{player_urls}{player url(s) from transfermarkt}
}
\value{
returns a dataframe of player injury history
}
\description{
Returns data frame of a player's injury history transfermarkt.com
}
