% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_transfer_history.R
\name{tm_player_transfer_history}
\alias{tm_player_transfer_history}
\title{Get Transfermarkt player transfer history}
\usage{
tm_player_transfer_history(player_urls, get_extra_info = TRUE)
}
\arguments{
\item{player_urls}{the player url(s) from transfermarkt}

\item{get_extra_info}{allows users to decide if they want to scrape extra info (contract length, countries involved) or not}
}
\value{
returns a dataframe of player transfers
}
\description{
Returns data frame of player(s) transfer history from transfermarkt.com
Replaces the deprecated function player_transfer_history
}
