% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_staff_job_history.R
\name{tm_staff_job_history}
\alias{tm_staff_job_history}
\title{Get Staff Member's job history}
\usage{
tm_staff_job_history(staff_urls)
}
\arguments{
\item{staff_urls}{transfermarkt.com staff(s) url (can use tm_league_staff_urls() to get)}
}
\value{
returns a data frame of all roles a selected staff member(s) has held and performance data
}
\description{
Returns all roles a selected staff member(s) has held and performance data
}
