% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldfootballr_helpers.R
\name{tm_team_player_urls}
\alias{tm_team_player_urls}
\title{Get transfermarkt Player URLs}
\usage{
tm_team_player_urls(team_url)
}
\arguments{
\item{team_url}{the player's team URL (can be from tm_league_team_urls())}
}
\value{
returns a character vector of all transfermarkt player URLs for a selected team
}
\description{
Returns the transfermarkt URLs for all players for a given team
}
