% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_team_staff_history.R
\name{tm_team_staff_history}
\alias{tm_team_staff_history}
\title{Get team staff history}
\usage{
tm_team_staff_history(team_urls, staff_role = "Manager")
}
\arguments{
\item{team_urls}{transfermarkt.com team(s) url for a season}

\item{staff_role}{the role description which can be found here:
https://github.com/JaseZiv/worldfootballR_data/blob/master/raw-data/transfermarkt_staff/tm_staff_types.csv}
}
\value{
returns a data frame of all selected staff roles for a team(s) history
}
\description{
Returns all people who have held the selected role in a team's history
}
