% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldfootballr_helpers.R
\name{tm_team_staff_urls}
\alias{tm_team_staff_urls}
\title{Get transfermarkt Club Staff URLs}
\usage{
tm_team_staff_urls(team_urls, staff_role)
}
\arguments{
\item{team_urls}{the staff member's team URL (can be from tm_league_team_urls())}

\item{staff_role}{role of the staff member URLs required for with options including:

\emph{"Manager"}, \emph{"Assistant Manager"}, \emph{"Goalkeeping Coach"},
\emph{"Fitness Coach"}, \emph{"Conditioning Coach"}}
}
\value{
returns a character vector of all transfermarkt staff URLs for a selected team(s)
}
\description{
Returns the transfermarkt URLs for all staff of selected roles for a given team
}
