% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_team_transfer_balances.R
\name{tm_team_transfer_balances}
\alias{tm_team_transfer_balances}
\title{Team transfer balances}
\usage{
tm_team_transfer_balances(country_name, start_year, league_url = NA)
}
\arguments{
\item{country_name}{the country of the league's players}

\item{start_year}{the start year of the season (2020 for the 20/21 season)}

\item{league_url}{league url from transfermarkt.com. To be used when country_name not avalilable in main function}
}
\value{
returns a dataframe of the summarised financial transfer performance of all teams for a league season
}
\description{
Returns all team's transfer aggregated performances for a chosen league season
}
