% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_team_transfers.R
\name{tm_team_transfers}
\alias{tm_team_transfers}
\title{Get team transfers}
\usage{
tm_team_transfers(team_url, transfer_window = "all")
}
\arguments{
\item{team_url}{transfermarkt.com team url for a season}

\item{transfer_window}{which window the transfer occurred - options include "all" for both, "summer" or "winter"}
}
\value{
returns a dataframe of all team transfers
}
\description{
Returns all transfer arrivals and departures for a given team season
}
