% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldfootballr_helpers.R
\name{understat_available_teams}
\alias{understat_available_teams}
\title{Get Understat available teams}
\usage{
understat_available_teams(league)
}
\arguments{
\item{league}{the available league in Understat as outlined below
#'
The leagues currently available for Understat are:
\emph{"EPL"}, \emph{"La liga}", \emph{"Bundesliga"},
\emph{"Serie A"}, \emph{"Ligue 1"}, \emph{"RFPL"}

#' @return a character vector teams names}
}
\description{
Returns all available team names for the selected league
}
\examples{
\dontrun{
try({
understat_available_teams(league = 'EPL')
})
}
}
