% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_match_results.R
\name{understat_league_match_results}
\alias{understat_league_match_results}
\title{Get Understat season match results}
\usage{
understat_league_match_results(league, season_start_year)
}
\arguments{
\item{league}{the available leagues in Understat as outlined below}

\item{season_start_year}{the year the season started

The leagues currently available for Understat are:
\emph{"EPL"}, \emph{"La liga}", \emph{"Bundesliga"},
\emph{"Serie A"}, \emph{"Ligue 1"}, \emph{"RFPL"}}
}
\value{
returns a dataframe of match results for a selected league season
}
\description{
Returns match results for all matches played in the selected league season from Understat.com
}
