% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_shots.R
\name{understat_match_shots}
\alias{understat_match_shots}
\title{Get Understat match shot locations}
\usage{
understat_match_shots(match_url)
}
\arguments{
\item{match_url}{the URL of the match played}
}
\value{
returns a dataframe of shooting locations for a selected team season
}
\description{
Returns shooting locations for a selected match from Understat.com
}
\examples{
\dontrun{
understat_match_shots(match_url = "https://understat.com/match/14789")
}
}
