% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_match_stats.R
\name{understat_match_stats}
\alias{understat_match_stats}
\title{Get Understat match stats table data}
\usage{
understat_match_stats(match_url)
}
\arguments{
\item{match_url}{the URL of the match played}
}
\value{
returns a dataframe with data from the stats table for the match
}
\description{
Returns the Stats values for a selected match from Understat.com
}
