% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_shots.R
\name{understat_player_shots}
\alias{understat_player_shots}
\title{Get all Understat shot locations for a player}
\usage{
understat_player_shots(player_url)
}
\arguments{
\item{player_url}{the URL of a selected player}
}
\value{
returns a dataframe of shooting locations for a selected player
}
\description{
Returns shooting locations for a selected player for all matches played from Understat.com
}
\examples{
\dontrun{
sterling <- understat_player_shots(player_url = "https://understat.com/player/618")
}
}
