% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldfootballr_helpers.R
\name{understat_team_meta}
\alias{understat_team_meta}
\title{Get Understat team info}
\usage{
understat_team_meta(team_names)
}
\arguments{
\item{team_names}{a vector of team names (can be just 1)}
}
\value{
a data.frame
}
\description{
Retrieve Understat team metadata, including team URLs. Similar to `understatr::get_team_meta`.
}
\examples{
\dontrun{
understat_team_meta(team_name = c("Liverpool", "Manchester City"))
}
}
