% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_teams.R
\name{understat_team_players_stats}
\alias{understat_team_players_stats}
\title{Get Understat team player stats}
\usage{
understat_team_players_stats(team_url)
}
\arguments{
\item{team_url}{the URL of the team season}
}
\value{
a dataframe of player stats for a selected team season
}
\description{
Retrieve Understat team player stats.
}
