% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_shots.R
\name{understat_team_season_shots}
\alias{understat_team_season_shots}
\title{Get Understat team season shot locations}
\usage{
understat_team_season_shots(team_url)
}
\arguments{
\item{team_url}{the URL of the team season}
}
\value{
returns a dataframe of shooting locations for a selected team season
}
\description{
Returns shooting locations for all matches played by a selected team from Understat.com
}
\examples{
\dontrun{
understat_team_season_shots(team_url = "https://understat.com/team/Manchester_City/2020")
}
}
