% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_teams.R
\name{understat_team_stats_breakdown}
\alias{understat_team_stats_breakdown}
\title{Get Understat team statistics breakdowns}
\usage{
understat_team_stats_breakdown(team_urls)
}
\arguments{
\item{team_urls}{the url(s) of the teams in question}
}
\value{
returns a dataframe of all stat groups and values
}
\description{
Returns a data frame for the selected team(s) with stats broken down in
different ways. Breakdown groups include:
}
\details{
\emph{"Situation"}, \emph{"Formation}", \emph{"Game state"},
\emph{"Timing"}, \emph{"Shot zones"}, \emph{"Attack speed"}, \emph{"Result"}
}
\examples{
\dontrun{
team_urls <- c("https://understat.com/team/Liverpool/2020",
              "https://understat.com/team/Manchester_City/2020")
understat_team_stats_breakdown(urls = team_urls)
}
}
