% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_player_scouting.R
\name{fb_player_scouting_report}
\alias{fb_player_scouting_report}
\title{Get fbref Full Player Scouting Report}
\usage{
fb_player_scouting_report(player_url, pos_versus)
}
\arguments{
\item{player_url}{the URL of the player (can come from fb_player_urls())}

\item{pos_versus}{either "primary" or "secondary" as fbref offer comparisons against multiple positions}
}
\value{
returns a dataframe of a player's full scouting information
}
\description{
Returns the scouting report for a selected player
}
\examples{
\dontrun{
fb_player_scouting_report(player_url = "https://fbref.com/en/players/d70ce98e/Lionel-Messi",
pos_versus = "primary")
fb_player_scouting_report(player_url = "https://fbref.com/en/players/d70ce98e/Lionel-Messi",
pos_versus = "secondary")
}
}
