% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_player_stats.R
\name{fb_player_scouting_report}
\alias{fb_player_scouting_report}
\title{Get fbref Player Scouting Report}
\usage{
fb_player_scouting_report(player_url)
}
\arguments{
\item{player_url}{the URL of the player (can come from fb_player_urls())}
}
\value{
returns a dataframe of a players scouting information
}
\description{
Returns the scouting report for a selected player
}
\examples{
\dontrun{
fb_player_scouting_report("https://fbref.com/en/players/3bb7b8b4/Ederson")
}
}
