% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fotmob_leagues.R
\name{fotmob_get_league_tables}
\alias{fotmob_get_league_tables}
\title{Get standings from fotmob}
\usage{
fotmob_get_league_tables(country, league_name, league_id)
}
\arguments{
\item{country}{Three character country code. Can be one or multiple. If provided, `league_name` must also be provided (of the same length)}

\item{league_name}{League names. If provided, `country` must also be provided (of the same length)}

\item{league_id}{Fotmob ID for the league. Only used if `country` and `league_name` are not specified.}
}
\value{
returns a dataframe of league standings
}
\description{
Returns league standings from fotmob.com. 3 types are returned: all, home, away
}
\examples{
\dontrun{
library(dplyr)
library(tidyr)

# one league
fotmob_get_league_tables(
  country = "ENG",
  league_name = "Premier League"
)

# one league, by id
fotmob_get_league_tables(
  league_id = 47
)

# multiple leagues (could also use ids)
league_tables <- fotmob_get_league_tables(
  country =     c("ENG",            "ESP"   ),
  league_name = c("Premier League", "LaLiga")
)

# look at tables if only away matches are considered
league_tables \%>\%
  dplyr::filter(table_type == "away")
}
}
