% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fotmob_stats.R
\name{fotmob_get_season_stats}
\alias{fotmob_get_season_stats}
\title{Get season statistics from fotmob}
\usage{
fotmob_get_season_stats(
  country,
  league_name,
  league_id,
  season_name,
  team_or_player = c("team", "player"),
  stat_name,
  stat_league_name = league_name,
  cached = TRUE
)
}
\arguments{
\item{country}{Three character country code. Can be one or multiple. If provided, `league_name` must also be provided (of the same length)}

\item{league_name}{League names. If provided, `country` must also be provided (of the same length).}

\item{league_id}{Fotmob ID for the league. Only used if `country` and `league_name` are not specified.}

\item{season_name}{Season names in the format `"2021/2022"`. Multiple allowed. If multiple leagues are specified, season stats are retrieved for each league.}

\item{team_or_player}{return statistics for either \code{"team"} or \code{"player"}. Can only be one or the other.}

\item{stat_name}{the type of statistic. Can be more than one.
`stat_name` may be one of the following, although it may not be available for both \code{"team"} or \code{"player"}:
\itemize{
\item{"Accurate passes per match"}
\item{"Average possession"}
\item{"Assists}
\item{"Big chances created"}
\item{"Big chances missed"}
\item{"Clean sheets"}
\item{"Clearances per match"}
\item{"Expected Goals"}
\item{"Expected Goals conceded"}
\item{"FotMob rating"}
\item{"Goals"}
\item{"Goals conceded per match"}
\item{"Goals per match"}
\item{"Key passes per match"}
\item{"Minutes per goal"}
\item{"Penalties awarded"}
\item{"Penalties conceded"}
\item{"Possession won final 3rd"}
\item{"Red cards"}
\item{"Saves per match"}
\item{"Shots on target per match"}
\item{"Successful dribbles per match"}
\item{"Successful tackles per match"}
\item{"Yellow cards"}
}}

\item{stat_league_name}{Same format as `league_name`. If not provided explicitly, then it takes on the same value as `league_name`. If provided explicitly, should be of the same length as `league_name` (or `league_id` if `league_name` is not provided).

Note that not Fotmob currently only goes back as far as `"2016/2017"`. Some leagues may not have data for that far back.}

\item{cached}{Whether to load the dataframe from the \href{https://github.com/JaseZiv/worldfootballR_data/blob/master/raw-data/fotmob-leagues/all_leagues.csv}{data CSV}. This is faster and most likely what you want to do, unless you identify a league that's being tracked by fotmob that's not in this pre-saved CSV.}
}
\value{
returns a dataframe of team or player stats
}
\description{
Returns team or player season-long statistics standings from fotmob.com.
}
\examples{
\dontrun{
epl_team_xg_2021 <- fotmob_get_season_stats(
  country = "ENG",
  league_name = "Premier League",
  season = "2020/2021",
  stat_type = "xg",
  team_or_player = "team"
)

## fotmob doesn't has data for 2016/2017 for some leagues, but not all.
##   you'll see a warning about this and receive an empty dataframe
get_epl_season_stats(
  season = "2016/2017"
)

epl_player_xg_2021 <- get_epl_season_stats(
  country = "ENG",
  league_name = "Premier League",
  season = "2020/2021",
  stat_type = "xg",
  team_or_player = "player"
)
}
}
