% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_lineups.R
\name{get_match_lineups}
\alias{get_match_lineups}
\title{Get match lineups}
\usage{
get_match_lineups(match_url, time_pause = 2)
}
\arguments{
\item{match_url}{the fbref.com URL for the required match}

\item{time_pause}{the wait time (in seconds) between page loads}
}
\value{
returns a dataframe with the team lineups for a selected match
}
\description{
Returns lineups for home and away teams for a selected match
}
\examples{
\dontrun{
match <- get_match_urls(country = "AUS", gender = "F", season_end_year = 2021, tier = "1st")[1]
df <- get_match_lineups(match_url = match)
}
}
