% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_shooting.R
\name{get_match_shooting}
\alias{get_match_shooting}
\title{Get match shooting event data}
\usage{
get_match_shooting(match_url, time_pause = 2)
}
\arguments{
\item{match_url}{the fbref.com URL for the required match}

\item{time_pause}{the wait time (in seconds) between page loads}
}
\value{
returns a dataframe
}
\description{
Returns detailed player shooting data for home and away teams for a selected match(es)
}
\examples{
\dontrun{
match <- "https://fbref.com/en/matches/bf52349b/Fulham-Arsenal-September-12-2020-Premier-League"
df <- get_match_shooting(match_url = match)
}

}
