% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_summary.R
\name{get_match_summary}
\alias{get_match_summary}
\title{Get match summary}
\usage{
get_match_summary(match_url)
}
\arguments{
\item{match_url}{the fbref.com URL for the required match}
}
\value{
returns a dataframe with the match events (goals, cards, subs) for selecetd matches
}
\description{
Returns match summary data for selected match URLs, including goals, subs and cards
}
\examples{
\dontrun{
match <- get_match_urls(country = "AUS", gender = "F", season_end_year = 2021, tier = "1st")
get_match_summary(match_url = match)
}

}
