% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_urls.R
\name{get_match_urls}
\alias{get_match_urls}
\title{Get match URLs}
\usage{
get_match_urls(country, gender, season)
}
\arguments{
\item{country}{the three character country code}

\item{gender}{gender of competition, either "M" or "F"}

\item{season}{the year the season concludes, in quotes, ie "2021"}
}
\value{
returns a character vector of all fbref match URLs for a given competition, season and gender
}
\description{
Returns the URL for each match played for a given league season
}
\examples{
\dontrun{
get_match_urls(country = "ENG", gender = "M", season_end_year = "2020")
}
}
