% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_squad_stats.R
\name{tm_squad_stats}
\alias{tm_squad_stats}
\title{Get squad player stats}
\usage{
tm_squad_stats(team_url)
}
\arguments{
\item{team_url}{transfermarkt.com team url for a season}
}
\value{
returns a dataframe of all player stats for team(s)
}
\description{
Returns basic stats for players of a team season
}
\examples{
\dontrun{

bayern <- tm_squad_stats(
team_url = "https://www.transfermarkt.com/fc-bayern-munchen/startseite/verein/27/saison_id/2020"
)
# can even do it for a number of teams:
team_urls <- tm_league_team_urls(country_name = "England", start_year = 2020)
epl_team_players_2020 <- tm_squad_stats(team_url = team_urls)
}
}
