% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_team_transfers.R
\name{tm_team_transfers}
\alias{tm_team_transfers}
\title{Get team transfers}
\usage{
tm_team_transfers(team_url)
}
\arguments{
\item{team_url}{transfermarkt.com team url for a season}
}
\value{
returns a dataframe of all team transfers
}
\description{
Returns all transfer arrivals and departures for a given team season
}
\examples{
\dontrun{

bayern <- tm_team_transfers(
team_url = "https://www.transfermarkt.com/fc-bayern-munchen/startseite/verein/27/saison_id/2020"
)
# can even do it for a number of teams:
team_urls <- tm_league_team_urls(country_name = "England", start_year = 2020)
epl_xfers_2020 <- tm_team_transfers(team_url = team_urls)
}
}
