% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_match_players.R
\name{understat_match_players}
\alias{understat_match_players}
\title{Get Understat match player data}
\usage{
understat_match_players(match_url)
}
\arguments{
\item{match_url}{the URL of the match played}
}
\value{
returns a dataframe with data for all players for the match
}
\description{
Returns player values for a selected match from Understat.com
}
