% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_match_stats.R
\name{understat_match_stats}
\alias{understat_match_stats}
\title{Get all Understat shot locations for a player}
\usage{
understat_match_stats(match_url)
}
\arguments{
\item{player_url}{the URL of a selected player}
}
\value{
returns a dataframe of shooting locations for a selected player
}
\description{
Returns shooting locations for a selected player for all matches played from Understat.com
}
