% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understat_teams.R
\name{understat_team_players_stats}
\alias{understat_team_players_stats}
\title{Get Understat team player stats}
\usage{
understat_team_players_stats(team_names, years)
}
\arguments{
\item{team_names}{a vector of team names (can be just 1)}

\item{years}{a vector of years/seasons (can be just 1)}
}
\value{
a tibble
}
\description{
Retrieve Understat team player stats. Similar to `understatr::get_team_players_stats`.
}
\examples{
\dontrun{
understat_team_players_stats(team_names = c("Liverpool", "Manchester City"), years = c(2019, 2020))
}
}
